Create setchargingFunction
This is a function to turn on conservation mode on Lenovo IdeaPad on Linux. 
To use it, first put the function in the ".bashrc" file and then call the function "set charging x".
here x can be either 1 (conservation mode on) or 0 (conservation mode off). if x is not passed, 1 is taken by default.